Separate extensions into new file.
