+++ lib/src/adapters/sqlite/moor_models.dart
@@ -4,6 +4,7 @@ import 'package:moor/ffi.dart';
 import 'package:moor/moor.dart';
 import 'package:numerote_core/src/models/label.dart' as core;
 import 'package:numerote_core/src/models/note.dart' as core;
+import 'package:numerote_core/src/adapters/sqlite/moor_extensions.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart' as p;
 part 'moor_models.g.dart';
@@ -42,6 +43,7 @@ class NoteWithLabels {
 LazyDatabase _openConnection() {
   return LazyDatabase(() async {
     final dbFolder = await getApplicationDocumentsDirectory();
+    // TODO: Decide on naming for DB
     final file = File(p.join(dbFolder.path, 'db.sqlite'));
     return VmDatabase(file);
   });
@@ -141,78 +143,5 @@ class MoorDatabase extends _$MoorDatabase {
       await (delete(labels)).go();
       await (delete(notes)).go();
     });
-  }
-}
-
-extension MoorLabelExt on Label {
-  core.Label toCoreLabel() {
-    return core.Label(
-      documentId: documentId,
-      name: name,
-      createdAt: createdAt,
-    );
-  }
-}
-
-extension _LabelExt on core.Label {
-  LabelsCompanion toCompanion() {
-    return LabelsCompanion.insert(
-      documentId: documentId,
-      name: name,
-      createdAt: createdAt,
-    );
-  }
-}
-
-extension _NoteText on core.Note {
-  NotesCompanion toCompanion() {
-    return NotesCompanion.insert(
-      documentId: documentId,
-      contents: contents,
-      createdAt: createdAt,
-      updatedAt: updatedAt,
-    );
-  }
-}
-
-extension MoorDatabaseExt on _$MoorDatabase {
-  Future<List<NoteWithLabels>> populateLabels(List<Note> notes) async {
-    final notesMap = {for (final note in notes) note.documentId: note};
-    final rows = await (select(noteEntries).join(
-      [
-        innerJoin(
-          labels,
-          labels.documentId.equalsExp(noteEntries.label),
-        ),
-      ],
-    )..where((noteEntries.note.isIn(notesMap.keys))))
-        .get();
-
-    final labelListMap = <String, List<Label>>{};
-    for (final row in rows) {
-      final label = row.readTable(labels);
-      final documentId = row.readTable(noteEntries).note;
-      labelListMap.putIfAbsent(documentId, () => []).add(label);
-    }
-
-    return notesMap.keys
-        .where((id) => notesMap[id] != null)
-        .map((id) => NoteWithLabels(
-              note: notesMap[id]!,
-              labels: labelListMap[id] ?? [],
-            ))
-        .toList();
-  }
-}
-
-extension MoorNoteLabelsExtension on NoteWithLabels {
-  core.Note toCoreNote() {
-    return core.Note(
-      documentId: note.documentId,
-      contents: note.contents,
-      labels: labels.map((e) => e.toCoreLabel()).toList(),
-      createdAt: note.createdAt,
-      updatedAt: note.updatedAt,
-    );
   }
 }
